Remove redundant re-export from server.ts
